+++ src/Runtime/mmdCamera.ts
@@ -97,7 +97,7 @@ export class MmdCamera extends Camera implements IMmdCamera {
 
     /**
      * Destroy a runtime animation by its handle
-     * * @param handle The handle of the runtime animation to destroy
+     * @param handle The handle of the runtime animation to destroy
      * @returns True if the animation was destroyed, false if it was not found
      */
     public destroyRuntimeAnimation(handle: MmdRuntimeAnimationHandle): boolean {